polars_chat_bot/requirements.txt
langchain
langchain-pinecone
python-dotenv
pinecone-client
beautifulsoup4
langchain-community
unstructured
boto3
langchain_aws